requirements.txt
# Google Cloud AI Platform SDK for using Vertex AI models
google-cloud-aiplatform

# Google Cloud Text-to-Speech API
google-cloud-texttospeech

# Vertex AI SDK for accessing specific models
vertexai

# LangChain for working with document loaders, retrievers, and LLMs
vertexai

# Google Generative AI Embeddings for vector generation
langchain_google_genai

# PyPDFLoader for loading PDFs into LangChain
langchain_community[texttospeech]

langchain_google_community

pydantic-settings

moviepy==2.0.0.dev2

google-cloud-vision

streamlit>=1.24.0

fastapi>=0.68.0

uvicorn>=0.15.0

python-dotenv>=0.19.0

langchain>=0.3.13

google-cloud-aiplatform>=1.25.0

google-generativeai>=0.2.0

pydantic>=1.8.0

python-multipart>=0.0.5
